% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ettj_tbl}
\alias{ettj_tbl}
\title{US yield curve data}
\format{
8463 rows and 13 columns.
}
\source{
\url{https://home.treasury.gov/policy-issues/financing-the-government/interest-rate-statistics}
}
\usage{
ettj_tbl
}
\description{
A \code{tibble} with constant maturity rates of the US yield curve.
}
\keyword{datasets}
