% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_5f.R
\name{factors_5f}
\alias{factors_5f}
\title{Estimates the 5 factors and 3 decay parameters in a Svensson (1994) like factor model. The decay parameters are optimized using the two step approach.}
\usage{
factors_5f(lambda, yields, maturidades)
}
\arguments{
\item{lambda}{Initial guess for the 3 decay parameters. Numeric values.}

\item{yields}{Observed yields. Numeric Vector.}

\item{maturidades}{Time to maturity. Numeric. Must match the length of Y and be in the same base, i.e, annual, monthly etc.}
}
\value{
A matrix containing the 5 factors and 3 decay parameters, respectively.
}
\description{
Determines the optimal decay parameters calling the rmse function.
Then, performs an OLS regression using the 5 factor yield curve model in order to determine the optimal \eqn{\lambda} decay parameters:
\deqn{y_{\tau} = \beta_{1} +
                 \beta_{2}((1-e^{(-\tau^{'}  \lambda_{1})})/(\tau^{'}\lambda_{1})) +
                 \beta_{3}((1-e^{(-\tau^{'}  \lambda_{1})})/(\tau^{'}\lambda_{1}) - e^{(-\tau^{'}  \lambda_{1})}) +
                 \beta_{4}((1-e^{(-\tau^{'}  \lambda_{2})})/(\tau^{'}\lambda_{2}) - e^{(-\tau^{'}  \lambda_{2})}) +
                 \beta_{5}((1-e^{(-\tau^{'}  \lambda_{3})})/(\tau^{'}\lambda_{3}) - e^{(-\tau^{'}  \lambda_{3})})}
}
\examples{

library(Quandl)

dados <- Quandl("USTREASURY/YIELD")
yields <- dados[1,2:ncol(dados)]
maturidades <- c(1/12, 2/12, 3/12, 6/12, 1, 2, 3, 5, 7, 10, 20, 30);
factors <- factors_5f(c(.9,.035,.035), yields, maturidades)
yc <- yc5f(factors[6:8], factors[1:5], maturidades)

}
