% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_dl.R
\name{factors_dl}
\alias{factors_dl}
\title{Estimates the 3 factors and decay parameter in the Diebold and Li (2006) factor model. The decay parameters are optimized using the two step approach.}
\usage{
factors_dl(lambda, yields, maturidades)
}
\arguments{
\item{lambda}{Initial guess for the decay parameter. A numeric value.}

\item{yields}{Observed yields. Numeric Vector.}

\item{maturidades}{Time to maturity. Numeric. Must match the length of Y and be in the same base, i.e, annual, monthly etc.}
}
\value{
A matrix containing the 3 factors and decay parameter, respectively.
}
\description{
Determines the optimal decay parameters calling the rmse function.
Then, performs an OLS regression using the 3 factor yield curve model in order to determine the optimal \eqn{\lambda} decay parameters:
\deqn{y_{\tau} = \beta_{1} + 
                 \beta_{2}((1-e^{(-\tau^{'}  \lambda_{1})})/(\tau^{'}\lambda_{1})) +
                 \beta_{3}((1-e^{(-\tau^{'}  \lambda_{1})})/(\tau^{'}\lambda_{1}) - e^{(-\tau^{'}  \lambda_{1})})}
}
\examples{

library('Quandl')

dados <- Quandl("USTREASURY/YIELD")
yields <- dados[1,2:ncol(dados)]
maturidades <- c(1/12, 2/12, 3/12, 6/12, 1, 2, 3, 5, 7, 10, 20, 30);
factors <- factors_dl(c(.9,.035), yields, maturidades)
yc <- ycdl(factors[1], factors[1:3], maturidades)

}
