% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors_sv.R
\name{factors_sv}
\alias{factors_sv}
\title{Estimates the 4 factors and 2 decay parameters in the Svensson (1994) factor model. The decay parameters are optimized using the two step approach.}
\usage{
factors_sv(lambda, yields, maturidades)
}
\arguments{
\item{lambda}{initial guess for the decay parameters}

\item{yields}{A vector of yields}

\item{maturidades}{A vector of maturities. Be sure to put them in the same base as yield, i.e., per annum, per month etc.}
}
\value{
A matrix containing the 4 factors ad 2 decay parameters in this order
}
\description{
Estimates the 4 factors and 2 decay parameters in the Svensson (1994) factor model. The decay parameters are optimized using the two step approach.
}
