% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse_sv.R
\name{rmse_sv}
\alias{rmse_sv}
\title{Estimates the squared root of the mean squared error (rmse) for an OLS regression of the Svensson (1994) factor model}
\usage{
rmse_sv(lambda, Y, tau)
}
\arguments{
\item{lambda}{Decay parameters}

\item{Y}{Observed yields}

\item{tau}{Time to maturity}
}
\value{
squared root of mean squared error
}
\description{
Performs an OLS regression using the 4 factor yield curve model:

\eqn{y = \beta_{1}((1-e^{(-\tau^{'}  \lambda_{1})})/(\tau^{'}\lambda_{1})) +
               \beta_{2}((1-e^{(-\tau^{'}  \lambda_{1})})/(\tau^{'}\lambda_{1}) - e^{(-\tau^{'}  \lambda_{1})}) +
               \beta_{3}((1-e^{(-\tau^{'}  \lambda_{2})})/(\tau^{'}\lambda_{2}) - e^{(-\tau^{'}  \lambda_{2})})}
}
