% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{`ettj_package_data`}
\alias{`ettj_package_data`}
\alias{ettj_package_data}
\title{US yield curve data}
\format{
## `ettj_package_data`
A data frame with constant maturity rates of the US yield curve.
}
\source{
<https://home.treasury.gov/policy-issues/financing-the-government/interest-rate-statistics>
}
\usage{
ettj_package_data
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
}
\keyword{datasets}
