% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yc5f.R
\name{yc5f}
\alias{yc5f}
\title{5 factor model formula}
\usage{
yc5f(lambda, factors, tau)
}
\arguments{
\item{lambda}{Decay parameters, associated with the slope, curvature 1 and curvature 2 loadings.Numeric values.}

\item{factors}{Estimated factors (see factors_5f). Numeric Vector.}

\item{tau}{Time to maturity. Numeric. Must match the length of Y and be in the same base, i.e, annual, monthly etc.}
}
\value{
Vector of yields corresponding to the maturities.
}
\description{
Returns the yields associated with a set of maturities, using decay parameters and factors with the 5 factor formula.
}
\examples{

#library('Quandl')

#dados <- Quandl("USTREASURY/YIELD")
dados <- ettj_package_data
yields <- dados[1,2:ncol(dados)]
maturidades <- c(1/12, 2/12, 3/12, 6/12, 1, 2, 3, 5, 7, 10, 20, 30);
factors <- factors_5f(c(.9, .035, .035), yields, maturidades)
yc <- yc5f(factors[6:8], factors[1:5], maturidades)
}
