% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ycsv.R
\name{ycsv}
\alias{ycsv}
\title{svensson (1994) 4 factor model formula}
\usage{
ycsv(lambda, factors, tau)
}
\arguments{
\item{lambda}{Decay parameters}

\item{factors}{The estimated factors Beta 1 through 4}

\item{tau}{Time to maturity}
}
\value{
vector of yields corresponding to the maturities
}
\description{
Returns the yields associated with a set of maturities, using decay parameters and factors with the Svensson (1994) formula
}
